frequency is only an int, not a float (does not change behavior)

+++ src/com/t_oster/liblasercut/drivers/LaosCutter.java
@@ -316,7 +316,7 @@ public class LaosCutter extends LaserCutter
     }
   }
   
-  private float currentFrequency = -1;
+  private int currentFrequency = -1;
   private void setFrequency(PrintStream out, int frequency)
   {
     if (currentFrequency != frequency)
